fix: Changing the global call of babel-node to the local call inside the project dir

+++ utils/copy.js
@@ -4,11 +4,13 @@ import colors from './console_color'
 //copy a file from .. to ..
 export default function copy(from, to) {
     //if the destination file is not exists
-    // if (!fs.existsSync(to))
+    if (!fs.existsSync(to))
         fs.copyFile(from, to, (err) => {
             if (err)
                 console.log(colors[0], err);
             else
                 console.log(colors[4], `copy file from ${from} to ${to}`);
         });
+    else
+        console.log(colors[2], 'The file is already exist');
 }
